# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 托管智训营正式学习 API
  version: 2.0.0
  description: |
    阶段2正式契约。Cookie 会话、动态 CSRF、用户隔离、学习顺序、路线状态、
    解锁、评分结论和主进度均由后端负责。公开接口永不返回正确答案、参考答案、
    Rubric、硬性必达项、评分关键词、内部 topic ID、模型名称或运行参数。
servers:
  - url: /
security:
  - sessionCookie: []
tags:
  - {name: auth}
  - {name: worlds}
  - {name: learning}
  - {name: attempts}
  - {name: remediation}
  - {name: records}
paths:
  /api/health:
    get:
      security: []
      responses:
        '200': {description: OK}
  /api/auth/csrf:
    get:
      security: []
      responses:
        '200':
          description: CSRF 令牌
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CsrfResponse'}
  /api/auth/login:
    post:
      security: []
      parameters: [{$ref: '#/components/parameters/CsrfHeader'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CurrentUser'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/auth/me:
    get:
      responses:
        '200':
          description: 当前用户
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CurrentUser'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/auth/logout:
    post:
      parameters: [{$ref: '#/components/parameters/CsrfHeader'}]
      responses:
        '204': {description: 会话已失效}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/worlds:
    get:
      tags: [worlds]
      summary: 三个学习世界入口聚合
      responses:
        '200':
          description: 清算和监督为 BUILDING；核算为 OPEN
          content:
            application/json:
              schema: {$ref: '#/components/schemas/WorldsResponse'}
              examples:
                default: {$ref: './examples/worlds-response.json'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/lines/{line}/map:
    get:
      tags: [worlds]
      summary: 条线连续地图整页聚合
      parameters:
        - {$ref: '#/components/parameters/Line'}
      responses:
        '200':
          description: 地图骨架、状态、解锁和进度
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MapResponse'}
              examples:
                accounting: {$ref: './examples/accounting-map-response.json'}
        '409': {$ref: '#/components/responses/BusinessConflict'}
  /api/routes/{routeId}:
    get:
      tags: [learning]
      summary: 路线概览
      parameters: [{$ref: '#/components/parameters/RouteId'}]
      responses:
        '200':
          description: 四环节状态和下一步
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RouteOverview'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/routes/{routeId}/steps/{stepType}:
    get:
      tags: [learning]
      summary: 按需加载单个学习环节
      description: BASIC_PRACTICE 不返回答案、解析或提示；COMPREHENSIVE_PRACTICE 不返回任何评分私有资产。
      parameters:
        - {$ref: '#/components/parameters/RouteId'}
        - {$ref: '#/components/parameters/StepType'}
      responses:
        '200':
          description: 统一环节外壳
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StepResponse'}
        '409': {$ref: '#/components/responses/BusinessConflict'}
  /api/routes/{routeId}/steps/{stepType}/complete:
    post:
      tags: [learning]
      summary: 知识卡或正常示范完成事件
      parameters:
        - {$ref: '#/components/parameters/RouteId'}
        - {$ref: '#/components/parameters/StepType'}
        - {$ref: '#/components/parameters/CsrfHeader'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CompleteStepRequest'}
      responses:
        '200':
          description: 幂等完成后的最新进度
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RouteProgress'}
        '409': {$ref: '#/components/responses/BusinessConflict'}
  /api/routes/{routeId}/basic-practice/{questionId}/answers:
    post:
      tags: [learning]
      summary: 基础练习逐题作答
      description: 首次响应仅返回正确与否、解释和分层提示，不返回答案键。
      parameters:
        - {$ref: '#/components/parameters/RouteId'}
        - {name: questionId, in: path, required: true, schema: {type: string}}
        - {$ref: '#/components/parameters/CsrfHeader'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PracticeAnswerRequest'}
      responses:
        '200':
          description: 即时反馈
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PracticeFeedback'}
              examples:
                wrong: {$ref: './examples/basic-practice-feedback.json'}
        '409': {$ref: '#/components/responses/BusinessConflict'}
  /api/routes/{routeId}/draft:
    get:
      tags: [learning]
      summary: 获取当前用户综合实务草稿
      parameters: [{$ref: '#/components/parameters/RouteId'}]
      responses:
        '200':
          description: 当前草稿
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DraftResponse'}
    put:
      tags: [learning]
      summary: 自动保存综合实务草稿
      parameters:
        - {$ref: '#/components/parameters/RouteId'}
        - {$ref: '#/components/parameters/CsrfHeader'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SaveDraftRequest'}
      responses:
        '200':
          description: 已保存
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DraftResponse'}
        '409': {$ref: '#/components/responses/BusinessConflict'}
  /api/routes/{routeId}/attempts:
    post:
      tags: [attempts]
      summary: 正式提交综合实务
      description: |
        在当前用户范围使用 clientRequestId 幂等。立即创建不可变答案、内容和 Rubric 快照，
        返回 attemptId；后台异步评分。重复请求必须复用同一标识。
      parameters:
        - {$ref: '#/components/parameters/RouteId'}
        - {$ref: '#/components/parameters/CsrfHeader'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SubmitAttemptRequest'}
            examples:
              default: {$ref: './examples/submit-attempt-request.json'}
      responses:
        '200':
          description: 已创建或返回同一正式作答
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AttemptResponse'}
        '409': {$ref: '#/components/responses/BusinessConflict'}
  /api/attempts/{attemptId}:
    get:
      tags: [attempts]
      summary: 查询异步评分状态与结果
      description: 查询只读取同一快照，绝不重新调用模型。
      parameters: [{$ref: '#/components/parameters/AttemptId'}]
      responses:
        '200':
          description: SCORING、COMPLETED 或 FAILED
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AttemptResponse'}
              examples:
                scoring: {$ref: './examples/attempt-scoring-response.json'}
                completed: {$ref: './examples/attempt-completed-response.json'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/attempts/{attemptId}/retry-scoring:
    post:
      tags: [attempts]
      summary: 对原 FAILED attempt 重试技术评分
      parameters:
        - {$ref: '#/components/parameters/AttemptId'}
        - {$ref: '#/components/parameters/CsrfHeader'}
      responses:
        '200':
          description: 原 attempt 回到 SCORING
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AttemptResponse'}
        '409': {$ref: '#/components/responses/BusinessConflict'}
  /api/attempts/{attemptId}/remediation:
    get:
      tags: [remediation]
      summary: 获取与未通过 attempt 绑定的定向补学
      parameters: [{$ref: '#/components/parameters/AttemptId'}]
      responses:
        '200':
          description: 补学目标和针对练习
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RemediationPlan'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/attempts/{attemptId}/remediation/{targetId}/answers:
    post:
      tags: [remediation]
      summary: 提交定向练习答案
      parameters:
        - {$ref: '#/components/parameters/AttemptId'}
        - {name: targetId, in: path, required: true, schema: {type: string}}
        - {$ref: '#/components/parameters/CsrfHeader'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RemediationAnswerRequest'}
      responses:
        '200':
          description: 后端判断目标是否完成
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RemediationFeedback'}
  /api/attempts/{attemptId}/comprehensive-practice-retry:
    post:
      tags: [remediation]
      summary: 全部补学完成后解锁综合实务重做
      parameters:
        - {$ref: '#/components/parameters/AttemptId'}
        - {$ref: '#/components/parameters/CsrfHeader'}
      responses:
        '200': {description: 综合实务已解锁}
        '409': {$ref: '#/components/responses/BusinessConflict'}
  /api/training-records:
    get:
      tags: [records]
      summary: 服务端分页训练记录
      parameters:
        - {name: page, in: query, schema: {type: integer, minimum: 0, default: 0}}
        - {name: size, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
        - {name: line, in: query, schema: {$ref: '#/components/schemas/Line'}}
        - {name: conclusion, in: query, schema: {$ref: '#/components/schemas/Conclusion'}}
      responses:
        '200':
          description: 正式作答记录分页
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TrainingRecordPage'}
  /api/training-records/{attemptId}:
    get:
      tags: [records]
      summary: 不可变历史详情
      description: 同时返回历史结论和路线当前状态。
      parameters: [{$ref: '#/components/parameters/AttemptId'}]
      responses:
        '200':
          description: 历史快照
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AttemptResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
components:
  securitySchemes:
    sessionCookie: {type: apiKey, in: cookie, name: JSESSIONID}
  parameters:
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      schema: {type: string}
    Line:
      name: line
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Line'}
    RouteId:
      name: routeId
      in: path
      required: true
      schema: {type: string, example: ACC-LIFE-ROLE-001}
    StepType:
      name: stepType
      in: path
      required: true
      schema: {$ref: '#/components/schemas/StepType'}
    AttemptId:
      name: attemptId
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
  responses:
    BadRequest:
      description: 请求格式无效
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Unauthorized:
      description: 未登录或会话过期
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Forbidden:
      description: CSRF 或安全校验失败
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    NotFound:
      description: 资源不存在或不属于当前用户
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    BusinessConflict:
      description: 幂等冲突、版本冲突、学习顺序、补学或评分状态不允许
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
  schemas:
    Line: {type: string, enum: [CLEARING, ACCOUNTING, SUPERVISION]}
    Availability: {type: string, enum: [OPEN, BUILDING]}
    RouteState: {type: string, enum: [LOCKED, NOT_STARTED, IN_PROGRESS, LEARNED_NOT_MASTERED, PASSED]}
    StepType: {type: string, enum: [KNOWLEDGE_CARD, DEMONSTRATION, BASIC_PRACTICE, COMPREHENSIVE_PRACTICE]}
    ProcessingStatus: {type: string, enum: [SCORING, COMPLETED, FAILED]}
    Conclusion: {type: string, enum: [PASSED, LEARNED_NOT_MASTERED]}
    Dimension: {type: string, enum: [CONCEPT, PROCESS, RISK, EXPRESSION]}
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [BAD_REQUEST, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, INTERNAL_ERROR,
                 IDEMPOTENCY_CONFLICT, CONTENT_BUILDING, CONTENT_VERSION_MISMATCH,
                 LEARNING_SEQUENCE_VIOLATION, DRAFT_CONFLICT, ATTEMPT_SCORING,
                 SCORING_RETRY_REQUIRED, SCORING_RETRY_NOT_ALLOWED, REMEDIATION_REQUIRED]
        message: {type: string}
    CsrfResponse:
      type: object
      required: [token, headerName]
      properties: {token: {type: string}, headerName: {type: string}}
    LoginRequest:
      type: object
      additionalProperties: false
      required: [employeeNo, password]
      properties:
        employeeNo: {type: string, maxLength: 32}
        password: {type: string, format: password, writeOnly: true, maxLength: 128}
    CurrentUser:
      type: object
      required: [employeeNo, displayName]
      properties: {employeeNo: {type: string}, displayName: {type: string}}
    WorldsResponse:
      type: object
      required: [mapVersion, worlds]
      properties:
        mapVersion: {type: string}
        worlds:
          type: array
          minItems: 3
          maxItems: 3
          items:
            type: object
            required: [line, name, availability, passedRequiredRoutes, publishedRequiredRoutes, progressPercent, status]
            properties:
              line: {$ref: '#/components/schemas/Line'}
              name: {type: string}
              description: {type: string}
              availability: {$ref: '#/components/schemas/Availability'}
              sceneAssetId: {type: string}
              passedRequiredRoutes: {type: integer}
              publishedRequiredRoutes: {type: integer}
              progressPercent: {type: integer}
              status: {type: string}
    MapResponse:
      type: object
      description: regions/modules/nodes 保持地图层级；每个节点由后端返回 state、locked、enterable、completedSteps。
      additionalProperties: true
    RouteOverview:
      type: object
      required: [routeId, contentVersion, rubricVersion, title, state, steps, nextStep, completedSteps, totalSteps]
      additionalProperties: true
    StepResponse:
      type: object
      required: [routeId, contentVersion, stepType, content, completed]
      properties:
        routeId: {type: string}
        contentVersion: {type: string}
        stepType: {$ref: '#/components/schemas/StepType'}
        content:
          type: object
          description: BASIC_PRACTICE 的 content 为 BasicPracticeContent；题面只返回结构化输入所需字段，不返回 answer、explanation 或 hints。
        completed: {type: boolean}
    BasicPracticeContent:
      type: object
      additionalProperties: false
      required: [questions]
      properties:
        questions:
          type: array
          minItems: 1
          maxItems: 5
          items: {$ref: '#/components/schemas/BasicPracticeQuestion'}
    BasicPracticeQuestion:
      type: object
      additionalProperties: false
      required: [questionId, type, prompt]
      properties:
        questionId: {type: string}
        type:
          type: string
          enum: [SINGLE_CHOICE, MULTIPLE_CHOICE, ORDERING, FIELD_MAP, CALCULATION, LEDGER_ENTRY, RECONCILIATION, SHORT_TEXT]
        prompt: {type: string}
        options:
          type: array
          items: {$ref: '#/components/schemas/PracticeChoiceOption'}
        items:
          type: array
          items: {$ref: '#/components/schemas/PracticeOrderingItem'}
        fieldMappings:
          type: array
          items: {$ref: '#/components/schemas/PracticeFieldMapping'}
        calculation:
          type: object
          required: [fields]
          properties:
            fields:
              type: array
              items: {$ref: '#/components/schemas/PracticeCalculationField'}
        ledgerEntries:
          type: array
          items: {$ref: '#/components/schemas/PracticeLedgerEntry'}
        reconciliation:
          type: object
          required: [fields]
          properties:
            fields:
              type: array
              items: {$ref: '#/components/schemas/PracticeReconciliationField'}
        textInput:
          type: object
          required: [label, placeholder]
          properties:
            label: {type: string}
            placeholder: {type: string}
    PracticeChoiceOption:
      type: object
      required: [optionId, text]
      properties:
        optionId: {type: string}
        text: {type: string}
    PracticeOrderingItem:
      type: object
      required: [itemId, text]
      properties:
        itemId: {type: string}
        text: {type: string}
    PracticeFieldMapping:
      type: object
      required: [fieldId, label, source, options]
      properties:
        fieldId: {type: string}
        label: {type: string}
        source: {type: string}
        options:
          type: array
          items: {$ref: '#/components/schemas/PracticeChoiceOption'}
    PracticeCalculationField:
      type: object
      required: [fieldId, label, formula, placeholder, unit]
      properties:
        fieldId: {type: string}
        label: {type: string}
        formula: {type: string}
        placeholder: {type: string}
        unit: {type: string}
        precision:
          type: integer
          minimum: 0
          maximum: 8
          description: 建议显示的小数位数；不限制学员输入更多有效小数位。
        tolerance:
          type: number
          minimum: 0
          description: 数值判定允许的绝对误差；未提供时按精确数值匹配。
    PracticeLedgerEntry:
      type: object
      required: [entryId, direction, label, amount, placeholder]
      properties:
        entryId: {type: string}
        direction: {type: string}
        label: {type: string}
        amount: {type: string}
        placeholder: {type: string}
    PracticeReconciliationField:
      type: object
      required: [fieldId, label, kind, formula, placeholder]
      properties:
        fieldId: {type: string}
        label: {type: string}
        kind: {type: string, enum: [NUMBER, SELECT]}
        formula: {type: string}
        placeholder: {type: string}
        precision:
          type: integer
          minimum: 0
          maximum: 8
          description: 建议显示的小数位数；不限制学员输入更多有效小数位。
        tolerance:
          type: number
          minimum: 0
          description: 数值判定允许的绝对误差；未提供时按精确数值匹配。
        options:
          type: array
          items: {$ref: '#/components/schemas/PracticeChoiceOption'}
    CompleteStepRequest:
      type: object
      additionalProperties: false
      required: [eventId, contentVersion]
      properties:
        eventId: {$ref: '#/components/schemas/IdempotencyKey'}
        contentVersion: {type: string}
    IdempotencyKey:
      type: string
      minLength: 8
      maxLength: 64
      pattern: '^[A-Za-z0-9_-]{8,64}$'
    PracticeAnswerRequest:
      type: object
      additionalProperties: false
      required: [requestId, contentVersion, answer]
      properties:
        requestId: {$ref: '#/components/schemas/IdempotencyKey'}
        contentVersion: {type: string}
        answer:
          type: array
          minItems: 1
          items: {type: string}
          description: 按题型顺序提交字段映射、数字复算、凭证科目、勾稽字段或一段结论；服务端按题型匹配，题面答案不下发。
    PracticeFeedback:
      type: object
      required: [questionId, correct, correctOnce, explanation, practiceCompleted, progress]
      properties:
        questionId: {type: string}
        correct: {type: boolean}
        correctOnce: {type: boolean}
        explanation: {type: string}
        hint: {type: string}
        practiceCompleted: {type: boolean}
        progress: {$ref: '#/components/schemas/RouteProgress'}
    RouteProgress:
      type: object
      required: [routeId, state, completedSteps, totalSteps]
      properties:
        routeId: {type: string}
        state: {$ref: '#/components/schemas/RouteState'}
        completedSteps: {type: integer}
        totalSteps: {type: integer}
        nextStep: {$ref: '#/components/schemas/StepType'}
    SaveDraftRequest:
      type: object
      additionalProperties: false
      required: [contentVersion, answer]
      properties:
        contentVersion: {type: string}
        answer: {$ref: '#/components/schemas/ComprehensivePracticeAnswer'}
        expectedRevision: {type: integer, minimum: 0}
    DraftResponse:
      type: object
      required: [routeId, revision]
      properties:
        routeId: {type: string}
        contentVersion: {type: string}
        answer: {$ref: '#/components/schemas/ComprehensivePracticeAnswer'}
        revision: {type: integer}
        updatedAt: {type: string, format: date-time, nullable: true}
    SubmitAttemptRequest:
      type: object
      additionalProperties: false
      required: [clientRequestId, contentVersion, rubricVersion, answer]
      properties:
        clientRequestId: {$ref: '#/components/schemas/IdempotencyKey'}
        contentVersion: {type: string}
        rubricVersion: {type: string}
        answer: {$ref: '#/components/schemas/ComprehensivePracticeAnswer'}
    AttemptResponse:
      type: object
      required: [attemptId, routeId, processingStatus, submittedAt, contentVersion, rubricVersion]
      properties:
        attemptId: {type: integer, format: int64}
        routeId: {type: string}
        processingStatus: {$ref: '#/components/schemas/ProcessingStatus'}
        submittedAt: {type: string, format: date-time}
        contentVersion: {type: string}
        rubricVersion: {type: string}
        technicalErrorCode: {type: string}
        result: {$ref: '#/components/schemas/ScoringResult'}
        answerSnapshot: {$ref: '#/components/schemas/ComprehensivePracticeAnswer'}
        historicalConclusion: {$ref: '#/components/schemas/Conclusion'}
        currentRouteState: {$ref: '#/components/schemas/RouteState'}
        remediationSummary: {type: object}
        allowedActions: {type: array, items: {type: string}}
    ScoringResult:
      type: object
      description: 公开结果保留维度证据，但不返回 Rubric 项目、硬性必达项内容或内部项目 ID。
      required: [totalScore, passScore, conclusion, scoreThresholdMet, allMandatoryRequirementsMet, dimensions]
      properties:
        totalScore: {type: integer, minimum: 0, maximum: 100}
        passScore: {type: integer, example: 75}
        conclusion: {$ref: '#/components/schemas/Conclusion'}
        scoreThresholdMet: {type: boolean}
        allMandatoryRequirementsMet: {type: boolean}
        dimensions:
          type: array
          minItems: 4
          maxItems: 4
          items:
            type: object
            additionalProperties: false
            required: [dimension, score, maxScore, items]
            properties:
              dimension: {$ref: '#/components/schemas/Dimension'}
              score: {type: integer}
              maxScore: {type: integer}
              items:
                type: array
                items:
                  type: object
                  additionalProperties: false
                  required: [matched, evidence]
                  properties:
                    matched: {type: boolean}
                    evidence: {type: string}
    RemediationPlan:
      type: object
      required: [planId, attemptId, completedTargets, totalTargets, completed, practiceRetryUnlocked, targets]
      additionalProperties: true
    ComprehensivePracticeAnswer:
      type: object
      additionalProperties: false
      required: [responses]
      properties:
        responses:
          type: object
          minProperties: 1
          maxProperties: 100
          additionalProperties:
            oneOf:
              - {type: string, minLength: 1, maxLength: 500}
              - {type: number}
    RemediationAnswerRequest:
      type: object
      additionalProperties: false
      required: [requestId, answer]
      properties:
        requestId: {$ref: '#/components/schemas/IdempotencyKey'}
        answer: {type: array, minItems: 1, items: {type: string}}
    RemediationFeedback:
      type: object
      required: [targetId, correct, targetCompleted, planCompleted, explanation]
      properties:
        targetId: {type: string}
        correct: {type: boolean}
        targetCompleted: {type: boolean}
        planCompleted: {type: boolean}
        explanation: {type: string}
        hint: {type: string}
    TrainingRecordPage:
      type: object
      required: [items, page, size, totalElements, totalPages]
      properties:
        items: {type: array, items: {type: object}}
        page: {type: integer}
        size: {type: integer}
        totalElements: {type: integer, format: int64}
        totalPages: {type: integer}